Automatizacia-main/requirements.txt
flake8==6.0.0
PyYAML==6.0